Phase5_ProjectDevelopment/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-dotenv==1.0.1
bcrypt>=4.0.1
python-jose[cryptography]==3.3.0
groq==0.13.0
sentence-transformers==3.3.1
pymupdf==1.25.0
requests==2.32.3
feedparser==6.0.11
pydantic==2.10.3
pydantic-settings==2.7.0
chromadb>=1.5.0
httpx==0.28.1
python-multipart==0.0.20
numpy>=2.0.0
